requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
joblib>=1.1.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
imbalanced-learn>=0.10.0